✅ Fix Int test typing

+++ packages/core/src/Int.spec.ts
@@ -23,7 +23,7 @@ describe('Int', () => {
   });
   describeType({ describe, it, expect })(Int, {
     typeName: 'Int',
-    instances: () => [0, 1, 2, -1, minValue, maxValue],
+    instances: () => [0 as Int, 1 as Int, 2 as Int, -1 as Int, minValue as Int, maxValue as Int],
     notInstances: () => ['1', 1.1, undefined, minValue - 1, maxValue + 1],
   });
   describeCodec({ describe, it, expect })(Int, {
